requirements.txt
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.0
Flask==2.1.2
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.1
pymongo==4.1.1
python-dotenv==0.20.0
uuid==1.30
Werkzeug==2.1.2
WTForms==3.0.1
